+++ packages/sign_in_with_apple/sign_in_with_apple_platform_interface/lib/exceptions.dart
@@ -49,6 +49,16 @@ abstract class SignInWithAppleException implements Exception {
           code: AuthorizationErrorCode.failed,
           message: exception.message ?? 'no message provided',
         );
+      case 'authorization-error/credentialImport':
+        return SignInWithAppleAuthorizationException(
+          code: AuthorizationErrorCode.credentialImport,
+          message: exception.message ?? 'no message provided',
+        );
+      case 'authorization-error/credentialExport':
+        return SignInWithAppleAuthorizationException(
+          code: AuthorizationErrorCode.credentialExport,
+          message: exception.message ?? 'no message provided',
+        );
 
       case 'credentials-error':
         return SignInWithAppleCredentialsException(
@@ -114,6 +124,12 @@ enum AuthorizationErrorCode {
 
   /// The authorization attempt failed for an unknown reason.
   unknown,
+
+  /// The credential export request failed.
+  credentialExport,
+
+  /// The credential import request failed.
+  credentialImport,
 }
 
 /// A [SignInWithAppleException] indicating something went wrong while authenticating.